refactor(aicoding): move architect-rebind into the aicoding namespace (Phase 2)
The aicoding-creation specific endpoint
PUT /api/bots/{architect_bot_id}/architect-rebind moves out of the generic
bot_management package into adapters/http/aicoding + core/aicoding/services,
with its Protocol under api/aicoding. The /api/bots path is kept.

Production:
- app.py: mount architect_rebind_router
- di/aicoding_module: bind ArchitectRebindService + provide its Protocol
- bot_management/router: drop RebindArchitectRequest + endpoint
- api/bot_service: drop rebind_architect_bot(_batch) from Protocol
- core/bot_management/services/bot_service: drop the 4 rebind methods

Tests:
- mv test_bot_service_rebind_architect -> aicoding/services/test_architect_rebind_service
- new api/aicoding/test_architect_rebind_router (happy + error)
- drop TestRebindArchitectBot from bot_management test_router
- endpoint docstring: BotService -> ArchitectRebindService

Verified: service/router/bindings 173p; endpoint+arch 7p; coverage gate 16p.
