+++ src/backend/src/agentclaw/community/api/bot_service.py
@@ -50,8 +50,6 @@ class BotServiceProtocol(Protocol):
     def update_bot(self, *args: Any, **kwargs: Any) -> Any: ...
 
     def update_bot_ext(self, *args: Any, **kwargs: Any) -> Any: ...
-    def rebind_architect_bot(self, *args: Any, **kwargs: Any) -> Any: ...
-    def rebind_architect_bot_batch(self, *args: Any, **kwargs: Any) -> Any: ...
 
     def ensure_hosted_workspace(self, *args: Any, **kwargs: Any) -> Any: ...
